Updated page to rerender only the cart component when an item is deleted or added.

+++ server/migration.sql
@@ -4,5 +4,5 @@ CREATE TABLE items (
     id SERIAL PRIMARY KEY,
     name VARCHAR(255),
     style VARCHAR(255),
-    price DOUBLE
+    price INT
 );
